+++ src/ui/gui/statisticswindow.cpp
@@ -419,7 +419,7 @@ void StatisticsWindow::encodedVideoFrame(uint32_t size,
                                          float psnrY,
                                          float psnrU,
                                          float psnrV,
-                                         float networkLatencyMs,
+                                         int64_t networkLatencyMs,
                                          int64_t creationTimestamp)
 {
   updateValueBuffer(videoEncDelay_,
